executeScriptInFrame: Use height field for frameId
To allow frameId > 65534, the height field is also used.
It's rare to have so many frames, but in a single page application that
makes heavy use of frames, this number could eventually be exceeded.

So, let's increase the upper bound of frameId by also using the height
field of the image to transfer image information.

In both fields, the NULL-byte cannot be used, because the iamge size has
    to be non-zero.
The width field can carry 0xFFFE different values.
The height field can carry 0xFE different values.
So, the frameId's upper bound is 0xFFFE * 0xFE = 16 645 636.

If one creates a new frame every minute, then this frame number would
be exhausted after almost 32 years. It's probably safe to store the
frame number in 3 bytes.