worker_proxy: Fix tokens on event pages
Previously I assumed that the token can be generated only once, and
cached in sessionStorage and a local variable in the content script,
and assumed that the cached value in sessionStorage would be kept
until the last extension page/frame is removed.

This assumption is incorrect for event pages due to a bug in Chrome:
https://code.google.com/p/chromium/issues/detail?id=42599
(minimal test case: https://crbug.com/42599#c12).

To work around the problem, I no longer cache the token and generate
a new one when a token is requested after the background/event page
is loaded. This token is then guaranteed to be stable for the lifetime
of the background/event page. This is typically several seconds, which
is more than enough for authenticating messages between the content
script and extension frame:
1. Content script creates frame.
2. Content script requests token from background page.
3. Content script sends MessagePort with token to iframe.
4. Iframe requests token from background page.
5. Iframe verifies that the tokens match and accepts the MessagePort.

Because the event page won't suspend between step 2 and 4,
authentication will succeed for valid requests.